Refactors Dockerfile for enhanced security
Enhances the Dockerfile by adopting a multi-stage build process, improving security and reducing the final image size.

It moves the Singularity installation to a separate stage and copies the binary to a slim Debian base.

Also changes to running as a non-root user.

This isolates the application and reduces potential attack surface.